Shift build into container
The runner environment of ubuntu is severely neutered and attempting to
use the i386 repositories immediately causes dependency hell.

This resolves that by relying on the docker ubuntu 20.04 image.

In full, the changes are:

- Specify container to run in.
- Set environment variables `DEBIAN_FRONTEND` and `TZ`
Without this change, the package installation process will wait asking
for input forever (TZ because there's an indirect dependency on tzdata).
- Mount the location of the code (github.workspace) to /data in the
  container
- Alter references to github.workspace to reference the location it is
  mounted to in the container.